+++ tools/toolchain_profiler.py
@@ -20,22 +20,40 @@ EMPROFILE = int(os.getenv('EMPROFILE', '0'))
 
 
 class Logger(ContextDecorator):
+  depth = 0
+
   def __init__(self, name):
     self.name = name
 
+  def __call__(self, func):
+    if self.name is None:
+      self.name = func.__name__
+    return super().__call__(func)
+
   def __enter__(self):
+    if EMPROFILE == 2:
+      indentation = '  ' * Logger.depth
+      logger.info('%sstart block "%s"', indentation, self.name)
+      Logger.depth += 1
     self.start = time.time()
 
   def __exit__(self, exc_type, value, traceback):
     # When a block ends debug log the total duration.
     now = time.time()
+    duration = now - self.start
     if exc_type:
-      logger.debug('block "%s" raised an exception after %.2f seconds', self.name, now - self.start)
+      msg = 'block "%s" raised an exception after %.3f seconds'
     else:
-      logger.debug('block "%s" took %.2f seconds', self.name, now - self.start)
+      msg = 'block "%s" took %.3f seconds'
+    if EMPROFILE == 2:
+      Logger.depth -= 1
+      indentation = '  ' * Logger.depth
+      logger.info(indentation + msg, self.name, duration)
+    else:
+      logger.debug(msg, self.name, duration)
 
 
-if EMPROFILE:
+if EMPROFILE == 1:
   original_sys_exit = sys.exit
   original_subprocess_call = subprocess.call
   original_subprocess_check_call = subprocess.check_call
@@ -219,17 +237,21 @@ if EMPROFILE:
     class ProfileBlock(Logger):
       def __init__(self, block_name):
         super().__init__(block_name)
-        self.block_name = block_name
+        self.name = block_name
 
       def __enter__(self):
-        ToolchainProfiler.enter_block(self.block_name)
+        ToolchainProfiler.enter_block(self.name)
 
       def __exit__(self, type, value, traceback):
-        ToolchainProfiler.exit_block(self.block_name)
+        ToolchainProfiler.exit_block(self.name)
 
     @staticmethod
     def profile_block(block_name):
       return ToolchainProfiler.ProfileBlock(ToolchainProfiler.escape_string(block_name))
+
+    @staticmethod
+    def profile():
+      return ToolchainProfiler.ProfileBlock(None)
 
     @staticmethod
     def imaginary_pid():
@@ -250,3 +272,6 @@ else:
     @staticmethod
     def profile_block(block_name):
       return Logger(block_name)
+
+    def profile():
+      return Logger(None)